refactor(create_metric): move create_metric to utility module
- Remove the CreateMetric file
